+++ internal/server/handlers/ta_handler.go
@@ -2,6 +2,7 @@ package handlers
 
 import (
 	"github.com/gin-gonic/gin"
+	"s-portal/internal/domain/model"
 
 	// "s-portal/internal/domain/model"
 	"s-portal/internal/domain/service"
@@ -20,6 +21,11 @@ func NewTAHandler(TAService *service.TAService, authService *service.AuthService
 }
 
 func (h *TAHandler) GetClassTA(context *gin.Context) {
+	if !h.authService.AssertPermission(context) {
+		context.Status(401)
+		return
+	}
+
 	ta, err := h.TAService.GetTA(context)
 
 	if err != nil {
@@ -36,6 +42,11 @@ func (h *TAHandler) GetClassTA(context *gin.Context) {
 }
 
 func (h *TAHandler) GetClassTAByClassID(context *gin.Context) {
+	if !h.authService.AssertPermission(context) {
+		context.Status(401)
+		return
+	}
+
 	ta, err := h.TAService.GetTAByClassID(context)
 
 	if err != nil {
@@ -52,6 +63,11 @@ func (h *TAHandler) GetClassTAByClassID(context *gin.Context) {
 }
 
 func (h *TAHandler) CreateClassTA(context *gin.Context) {
+	if !h.authService.AssertPermission(context, model.RoleTeacher, model.RoleAdmin) {
+		context.Status(401)
+		return
+	}
+
 	err := h.TAService.CreateTA(context)
 
 	if err != nil {
@@ -68,6 +84,11 @@ func (h *TAHandler) CreateClassTA(context *gin.Context) {
 }
 
 func (h *TAHandler) UpdateClassTA(context *gin.Context) {
+	if !h.authService.AssertPermission(context, model.RoleTeacher, model.RoleAdmin) {
+		context.Status(401)
+		return
+	}
+
 	err := h.TAService.UpdateTA(context)
 
 	if err != nil {
@@ -84,6 +105,11 @@ func (h *TAHandler) UpdateClassTA(context *gin.Context) {
 }
 
 func (h *TAHandler) DeleteClassTA(context *gin.Context) {
+	if !h.authService.AssertPermission(context, model.RoleTeacher, model.RoleAdmin) {
+		context.Status(401)
+		return
+	}
+
 	err := h.TAService.DeleteTA(context)
 
 	if err != nil {
